refactor(financialrecommendations): move financial recommendations related hooks into seperate folder
